+++ sdk/redis/Azure.ResourceManager.Redis/CHANGELOG.md
@@ -10,6 +10,12 @@
 
 ### Other Changes
 
+## 1.0.1 (2022-09-27)
+
+### Bugs Fixed
+
+- Fixed the serialization issue of `RedisCommonConfiguration.RdbBackupMaxSnapshotCount`.
+
 ## 1.0.0 (2022-08-29)
 
 This is the first stable release of the Redis Management client library.